test(ducklake): move live-connection tests out of the fast unit lane
The unit module carries a module-level `fast` marker and
TestDuckLakeLiveConnection carried `live_integration`. Pytest markers
accumulate, so those 5 tests matched `fast and live_integration` and
tripped the CI fast-lane policy guard (FAST_LANE_VIOLATION).

Move the class to tests/integration/test_ducklake_integration.py, where
it belongs: no other tests/unit module carries live_integration. The
module-level INSTALL-ducklake probe moves with it, which also removes a
network extension install from fast-lane collection.

The unit module stays hermetic and mock-only.
